Update + Fix: Fixed image link errors in team page.js plus removed redundant files
